+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.65
-// Generated at: 2026-05-02T03:29:24.262Z
+// Spec version: 1.10.66
+// Generated at: 2026-05-02T03:33:01.020Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */